+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - Add ageband to statistic page !1685 @fs_k
 - Activity page shows only last entry to buddywall, eventwall, fairsharepointwall !1694 @fs_k
 - Leaving a district is captured in history. Leaving a home district removes verification. !1358 @fs_k
+- Add Button in message and bell list in the topbar to mark all bells and all conversations as readed #935 !1673 TheSoulT
 
 ## Changes
 - Removed FAQ section !1568 @chriswalg